2016 NATURAL APPROACH AS A PROPOSAL TO DEVELOP THE ORAL SKILL IN  FIFTH GRADE STUDENTS AT MANUEL RODRIGUEZ TORICES SCHOOL
ABSTRACT 
  This research project aims to develop the oral skill of students of fifth grade of Manuel 
Rodriguez Torices School through the implementation of the natural approach. This process is 
necessary because in this educational institution, a problem of low oral production in the English 
language was found in students mentioned above,  which was detected through observations in 
English classes, an interview to the English teacher in charge, an English diagnostic test and a 
survey were also administered  to 25 students from that grade. This process also shows how after 
the implementation of natural approach method the students were able to express their ideas 
according to the standards for fifth elementary grade. The present study is based on a paradigm of 
qualitative research and is a participatory action research which aims to fundamental social 
change, transform the reality and that people become aware of their responsibility in this process 
of transformation. 
Key words: natural approach, language learning, language acquisition