2016 DESIGNING A COMMUNICATIVE BOOKLET FOCUSED ON TASK BASED  LEARNING APPROACH  TO DEVELOP THE SPEAKING SKILL IN FOURTH  GRADE STUDENTS AT INSTITUCION EDUCATIVA ISABEL LA CATÓLICA
Abstract 
Currently, Colombia is a country that is interested in fostering the capacity of its 
citizens to learn a foreign language, because it is essential to construct a strong society about 
academic, cultural and economic aspects. The development of communicative competence 
levels in a foreign language of a society generate on its citizens more opportunities in 
relation to the aspects previously mentioned.  
In relation to the above, this research project represents a pedagogical support to the 
Colombia public education system, regarding the area of English as a foreign language, 
because it deals to develop the students’ English speaking skill through the design of 
communicative tasks that are focused on Task-based learning approach. 
This study takes place in the “Institución Educativa Isabel La Catolica”, where the 
proposal is necessary and appropriate to the institution because the English learning process 
in fourth grade presents difficulties regarding the methodology that is being implemented 
which is affecting the development of students’ speaking skill.  
This study was started when the students were in third grade and we have worked in 
three phases. Firstly, we had a period of observations with the intention of knowing the 
needs and interests of the studied population deeply, after that, we identified the problem 
situation and statement the problem in order to improve the students’ weaknesses and at the 
end a solution according to the evidenced problem situation was proposed.  
Key words: Task Based Learning Approach, speaking skill, development, education, 
teaching, learning, English, needs, interests.