2012 ESTRATEGIAS PARA ESTIMULAR LA COMPETENCIA  COMUNICATIVA (HABILIDAD ORAL) EN INGLÉS EN EL CURSO HIGH II  DEL COLEGIO BILINGÜE DE CARTAGENA
ABSTRACT 
 
This final report is about students in a course at private bilingual school (Bilingüe de 
Cartagena School in Colombia), who speak more Spanish than the foreign language in 
the English class, they are ten students (six girls and four boys), and they are between 14 
and 16. The students are not aware of the importance of English speaking inside and 
outside the classroom; they hardly ever pay attention to the teacher, because they always 
talk about different issues from the class, in the mother tongue. To check the 
communicative competence (oral skill) of students, they took an oral exam, it showed 
low percentages in some categories of the oral proficiency test. Therefore, the proposal 
for improvement included: conversational strategies, role plays, issues of interest (the 
institutional needs, ages and students wants); importance of speaking English outside 
and inside the classroom. The implementation of the proposal showed positive numbers, 
so it is possible to stimulate oral skills. Also, it was noticed that by conducting the 
classes with the communicative approach and topics of interest rather than focusing on 
aspects of grammar of the language, it makes the class more fun and interesting, 
therefore the participation of students is active. 
Keywords: communicative competence (oral skill), conversation strategies, role plays, 
bilingualism, teaching strategies