2016 DYNAMIC ASSESSMENT APPROACH IN ENGLISH LANGUAGE TEACHING AND LEARNING: A REVIEW
Abstract 
 
In the teaching and learning process, several forms of assessment has emerged in order to satisfy the need of evaluating student´s development. The most implemented form of assessment has been static assessment or better known as traditional assessment. As a response to change the way students have been evaluated, dynamic assessment appears as a form of assessing based on Vygotsky's sociocultural theory. This revision focuses on that new form of assessment, which is growing between the educators. This study presents an overview of dynamic assessment and its application in teaching foreign language, especially in the improvement of reading and writing skills. As a main result, the literature reveals that the implementation of DA in writing and reading skills is effective and it works as a functional complement of traditional assessments, but factors such as context, the amount of hours, teachers’ competences and training, are elements that have to be considered to assure the effectiveness of this approach. 
 
key words: Dynamic Assessment (DA), Static Assessment, Zone of proximal Development (ZPD), English as a Foreign language (EFL), English as a Second Language (SLL), Vigotsky`s socio cultural Theory.