2017 HOW TO DESIGN A PEDAGOGICAL FRAMEWORK TO ACHIEVE  ILLOCUTIONARY STRENGTH IN THE LEARNER'S ENGLISH EXPRESSIBILITY
Summary 
The act of speaking by the principle of expressiveness requires that there be analytic connections 
between the speech act (what the speaker means), what the phrase uttered means (or any other 
linguistic element), what the speaker intends (intention), what the listener understands (meaning) 
and the rules that govern the different linguistic elements, that must converge in the same plane 
of action so that the dynamics of the conversational circularity of a linguistic community is 
fulfilled. If this dynamics of circularity is not present in the conversational act, the "play of 
language" is not concretized. Therefore, it is important to point out that it is fundamental that the 
English teacher can assist the student's expressibility without interrupting the "degree" of 
discursive fluency in which the learner's linguistic resources and emotional resources allow the 
student to behave as a member of the new linguistic community. 
Keywords: language games, principle of expressibility, speech acts, social semiotics, 
conversation, discursive performance