2016 PORTFOLIO ASSESSMENT AS AN ALTERNATIVE INSTRUMENT TO  IMPROVE FOREIGN LANGUAGE LEARNING IN FOURTH AND FIFTH  STUDENTS ON A PRIVATE SCHOOL
Abstract 
          This study describes a research that was carried out in order to contribute to improve 
foreign language learning in fourth and fifth grade students of a private school.  It is so 
relevant to highlight that the assessment instrument could be used in order to promote, 
improve and facilitate meaningful learning. Data was collected through a participatory 
observation, structured interview and survey and using a type of research called action 
research. The results revealed that  the instruments of assessment used by the English 
teacher such as written exam, quiz and worksheet, these  did not contribute to achieve  the 
best outcomes in the English learning of the fourth and fifth graders but when the students 
did the assessments activities they had an active participation, they used the English  
language, they showed an interest in doing the activities and they did an self-reflection 
which allows them to realize that  they are improving their English learning. The results 
obtained with the students revealed that assessment activities as a preliminary phase of 
portfolio contributed in improving the English learning process of the students of fourth 
and fifth grade. They gained meaningful knowledge in the English language. 
      
Key words: portfolio assessment, learning, skills, English as a foreign language