2012 EVALUACIÓN DEL IDIOMA INGLÉS EN  LOS ESTUDIANTES DE BÁSICA SECUNDARIA Y MEDIA DEL COLEGIO BILINGÜE DE CARTAGENA
ABSTRACT 
 
Second language assessment has always been a harsh issue in the teaching and learning process. There are two important reasons for this consideration: The difficulty that testing developers face when designing exams that actually measure communicative competence and the unquestionable influence that these assessment instruments have on the methodological process in the classroom.  The main purpose of the present work was to contribute to the improvement of the evaluation system that would really measure the high school students’ communicative competence at Colegio Bilingue de Cartagena, Colombia. A diagnose of the current evaluation system being used at the institution was first done. This was done by observing a number of classes and evaluation sessions in the eighth grade and by giving a semi-structured interview that was administered to both the English teacher and the thirteen students in the eight grade. The information gathered was the starting point to develop assessment instruments based on the Bachman model (1990) of communicative language ability, the same model ICFES chose in order to design the new state examinations. Afterwards, the new exams were then trialed and administered again after having made a few necessary adjustments. Finally, a workshop was programmed for the English teachers at the Colegio Bilingüe, which was about all the aspecting of second language assessment as well as the communicative language ability model proposed by Bachman in 1990 so the schoolteachers could then have the theoretical framework to design their own exams.     Key words: assessment, second language, communicative competence, validity, reliability, washback.