2016 PROCESOS DE INTERACCIÓN EN EL AULA DE INGLÉS   Institución Educativa Técnica Agroindustrial de Bayunca 8 grado
ABSTRACT  
We can understand the interaction as an action exerted reciprocally between two or more 
subjects, objects, agents, forces or functions that involve cooperation between them in the 
educational field, which have similar affinities and goals that result of the acquisition of 
communication skills in a foreign language, due to they acted simultaneously on the same 
learning environments. Based on the above, this research addresses the factors affecting 
interactions in teaching-learning processes between teachers and students in the English subject, 
in a public institution of semi-urban area of the city of Cartagena de Indias.  
Key words: Interaction, semi-urban, English Language.