2014 PROPUESTA DE DISEÑO DE CURSO   PARA ÁREA DE INGLÉS EN QUINTO DE PRIMARIA DE INSTITUCIÓN EDUCATIVA PLAYAS DE ACAPULCO
ABSTRACT 
 
This project represents a pedagogical support to the public education system, regarding the area of English as foreign language; it takes place in the Institución Educativa Playas de Acapulco, where the proposal is necessary and appropriate to the institution due to it presents difficulties regarding the approaches and methods that guide the teaching and learning English processes. The proposal aims at the fifth grade. The exercise draws on classroom observations made to this group, since third grade that had a total of 35 children. The main goal of this study is contributing to strengthening of the teaching and learning English processes in the public schools Localidad 2 de la Virgen Cartagena, Bolivar. 
 
Key words: Curriculum; course design; syllabus; competences.