2018 ANXIETY IN ENGLISH LANGUAGE LEARNING: A CASE FROM COLOMBIA A Study of the Effect of Anxiety in Oral Production of Fifth Semester at Fundación Universitaria Colombo Internacional
Abstract
The purpose of the present research study was to determine the effects and causes of anxiety on students of Bachelor in Basic Education program with an emphasis in English in the Fundación Universitaria Colombo Internacional (UNICOLOMBO), making use of a mix research method to collect data in a population of 25 students of the fifth semester. Such data collection was analyzed through 3 categories that correspond to the categories exposed by Horwitz (1986) in his researches about anxiety. Great authors such as Horwitz, Krashen and Gardner were studied to establish the relationship between anxiety and English language learning processes. The data collection revealed impressive results that accurately describe the context and the situation UNICOLOMBO is having in terms of English learning processes. The main tool used to data collection was the Foreign Language Classroom Anxiety Scale (FLCAS), which has been made used to study Foreign Language Anxiety (FLA) around the world and to open new scenarios in Colombia about this issue.
Keywords: Foreign Language Anxiety, Foreign Language Classroom Anxiety Scale, English in Colombia