2016 ESTRATEGIAS TECNOLÓGICAS PARA EL FORTALECIMIENTO DE LA ESCUCHA EN EL APRENDIZAJE DEL INGLÉS DE LOS ESTUDIANTES DE 5º DE LA ESCUELA NORMAL SUPERIOR DE CARTAGENA DE INDIA
Abstract
This research focused on technological strategies for strengthening listening comprehensión in English to students in grades 5 to the Escuela Normal Superior in Cartagena de Indias. Its main objective is based on the promotion of technological strategies for strengthening listening comprehension .The research process was developed under a descriptive type with a methodology of qualitative-type and intervention action, whose function is to routinely investigate by learning innovative strategies to encourage students in their learning process. The problem of lack of listening was identified; it was evidenced through observations in the classroom, in a survey of students and teachers. To observe classes perceived that children do not understand what the teacher said to them in English, even for short dialogues. This perception was what motivated the design of technological strategies in pursuit of strengthening the ability of listening in English language.
The keywords that guided this project were: Strategies, Technology, Listening, Teaching and Learning.