2016 DEVELOPING SOCIOLINGUISTIC COMPETENCE IN FOREIGN LANGUAGE THROUGH ROLE PLAY ACTIVITIES WITH NINTH GRADE STUDENTS AT A PRIVATE SCHOOL OF CARTAGENA
ABSTRACT
This research project focuses on the sociolinguistic competence as an important
component in the teachinglearning
process of a foreign language. Even though it is recognized
that a language is a tool to exchange information and communicate with different people, many
times learners of a foreign language are unconscious or unaware of the different problems which
become a trouble when they are interacting in this language, that is the reason why it is
indispensable to mention some of the facts that make it difficult and crucial for this research,
register; formal, consultative and casual. These concepts are exposed because of the lack of
knowledge of learners when they are interacting in the target language. Besides, this study is
oriented under action research which took place in a seventeen ninth
grade learners’ group of a
private school in order to apply an improvement plan to give solutions to these problems.
Furthermore, this research emphasizes on the use of the second language to foster the
interactions in class with the purpose that students can identify, know and apply sociolinguistic
knowledge in order to be aware of which utterances are appropriate in the social situation or
context in which he or she is interacting. One of the main result of this study, the students were
able to use appropriate register in their English classes through role play activities and be aware
when use them in real situations.
Key words: sociolinguistic competence, EFL learning, language, interactions, language
register: formal, consultative and casual.