2016 IMPROVING THE SPEAKING SKILL THROUGH TWO COMMUNICATIVE ACTIVITIES: PICTURE DESCRIPTION AND INTERVIEWS ON STUDENTS FROM FOURTH GRADE AT INSTITUCIÓN EDUCATIVA OCTAVIANA DEL C. VIVES C.
ABSTRACT
In this document, the main idea is to use the picture description and interview activities in the teaching and learning process of students from fourth grade at Octaviana del C.Vives C. School to improve their speaking skill. This project is based on communicating through the target language; those activities are from communicative language teaching approach because this approach is one of the most appropriate in this context to use the language to communicate, due to allow students use the language to keep in touch with others and be active in the society.
Key Words: communicative language teaching approach, foreign language, speaking skill.