2018 DESCRIBING THE DEVELOPMENT OF COGNITIVE AND PHYSICAL SKILLS THROUGH VIDEO GAMES IN A SAMPLE OF THREE STUDENTS
ABSTRACT
Based on the amount of research, nowadays the education seems not to be quite efficient for students in terms of the needs that students have in each particular context. This perception is not only related to the teaching methods used nor the content being taught in class but also to the fact that the teaching strategies might not be the most appropriate for the students’ needs and wants. These two aspects are essential when designing or developing a syllabus in an institution, but unfortunately most stakeholders focus more on administration of non-educational elements such as incomes and management of the personal. When talking about needs and wants of the students, it is common to hear that videogames are one of the most recurrent answers from students as a way of learning English. Videogames is one of the most criticized ways of entertainment for students, especially for young learners because they are usually associated with distraction and waste of time. However it has been found that videogames can be also used as an educational tool. This is the main purpose of the following project, showing different cognitive and physical abilities that are developed through playing video games. This case study intends to describe and analyze different cognitive and physical abilities that gamers develop through this activity and how these abilities have allowed them to improve their in different aspects of their daily life. As a result, it has been proved that videogames are more than a simple distractor or entertainment gadget and that they can also be taken into account at the moment of promoting the physical and cognitive development in different types of population.
Key words: Video games, Cognitive skill, Physical skill, Achievements, Tool, Gamers.