2016 INTERCULTURALISM AND EDUCATION: GUIDELINES TO MATERIAL DESIGN FOR ENGLISH LEARNING IN FIFTH GRADE AT INSTITUCIÓN EDUCATIVA TÉCNICA DE PASACABALLOS.
Abstract
Teaching and learning process of English as a foreign language into cultural diversities
required to promote intercultural competences; in order to achieve that, teachers need to
design or adapt materials that recognize the student's own culture and the target culture to an
intercultural dialogue between the native and foreign culture. This research analyzes English
teaching and learning process in fifth grade at Institución Educativa Técnica de Pasacaballos,
an official and ethno educative school located in Cartagena de Indias, through a qualitative
research supported on the action research method that allowed to identify that cultural aspects
are not being taken into account into the English classroom, so it is proposed a guidelines to
material design through intercultural keys with the objective to improve English teaching
and learning recognizing cultural practices, beliefs, values and lifestyles of young learners
as rural and afrodescendant community members.
Key Words:
Interculturality, intercultural education, intercultural competence and intercultural material.