2014 DISEÑO DE IMPLEMENTACIÓN DE MATERIALES PARA LA ENSEÑANZA Y APRENDIZAJE DE INGLÉS EN EL GRADO QUINTO DE LA INSTITUCIÓN EDUCATIVA FOCO ROJO
ABSTRACT
Educational materials play an important role in the process of teaching and learning English because they help the teachers not only to lead the learning of the language of their students but also to interact and encourage them to keep on learning.Furthermore, the use of materials in class provides the learners with opportunities to learn meaningfully due to the fact that the materials open a path in which the students can link their previous knowledge with the new one and with reality.In Cartagena, the use of materials in the English classes is not taken as a priority; traditional classes are still in fashion, taking the board and the marker as the only resources to share information with the learners.At this point is where our research project takes a relevant position. The project was developed at InstituciónEducativaFocoRojo in which observations and interventions were made with the aim of providing the institution with suitable materials to improve the English level of its students, and consequently reaching the goal of the government with its ProgramaNacional de Inglés. The goal of this program is to offer Colombian citizens the necessary tools and spaces to improve their competences in the English language.
Keywords: didactic materials, teaching, learning, students.