2015 Enfoque ecológico en la enseñanza y aprendizaje del idioma inglés: una revisión
Abstract
Develops an approximation to the theme of the ecological approach in English
language teaching and learning, through a documentary review in database: avant, scielo,
eric, novitasroyal, dialnet, redalyc, educare, sciencedirect, eurojournal, obtaining a corpus
of 112 documents published during the last two decades, with the key words: ecological
approach, affordance, symbolic competence, interaction, multiculturalism, English as
foreign language and sociocultural
theory. Documents were included are in Spanish,
English and a text in Portuguese; titles and abstracts were reviewed separately and the
references were chosen and considered subjectively relevant and the ones allowed to meet
the objectives were reviewed completely. The results suggest that the ecological approach
in teaching and learning of a language involves a broad interdisciplinary network and is
built on strong sociocultural
principles, considering that language learning processes are
based on the interactions, adaptation and different factors of the environment, taking into
account the micro and macrosocial
systems as main pillars in the learning of a language.
Keyword: ecological approach, affordance, symbolic competence, interaction,
english as a foreign language, sociocultural theory.