2015 ESTRATEGIAS PEDAGÓGICAS BASADAS EN LOS ESTILOS COGNITIVOS DE LOS ESTUDIANTES DE 4°-2 DE LA INSTITUCIÒN EDUCATIVA CORAZON DE MARIA PARA LA DINAMIZACIÓN DEL PROCESO APRENDIZAJE- ENSEÑANZA DEL IDIOMA INGLÉS
Abstract
The interest in studying new ways of teaching English language has become increasingly indispensable in modern societies. For that reason, governments and educational institutions are concerned about the use of strategies and appropriate techniques that make a class more versatile and dynamic, taking into account not only how the language is taught but also how it is learned. In this way the purpose of our research offers a look into the learner, therefore the main objective is to analyze the incidence of the implementation of pedagogical strategies based on cognitive styles in a group of students from a public school at Cartagena city to encourage the learning-teaching process of english, aim fully achieved through techniques such as participant observation, interviews, workshops, among others, it was possible to understand the dynamic of the studied reality and propose the pedagogical strategy English Path which represents a new way of learning and teaching english in the studied sample, further it invites current and future language teachers to think of their educational practice regarding the particular way in which their students learn.
KEY WORDS: strategies, cognitive styles, learning-teaching, foreign language, stimulate.