2016 Teaching and Learning English through Specific ICTs Resources: a review
Abstract
This review paper states a close view of the most recent and used technologies for
English language teaching and learning, through a documentary review in database:
refseek, sciencedirect, academia and tandfonline, obtaining a corpus of 52 documents
published from 2005 and 2015, with the keywords: Teaching and Learning English
language, ICTs. The documents that were included are written in Spanish, English and a
text in Portuguese; all of the articles were written and considered relevant to this work as
they had a deep relation with the objectives of the paper. The results suggest that the use of
new technologies in teaching and learning English language foster and encourage students
and teachers in learning and adapting the language to any context, taking into consideration
why and how these technologies could make possible a successful and an enjoyable
learning process.
Keywords: Teaching and Learning English language, ICTs