2018 STRATEGIES USED IN THE ENGLISH LEARNING AND TEACHING PROCESS CONSIDERING CULTURE
Abstract
This article explores strategies for the teaching and learning of the English language
through a documentary review of approximately 35 documents published during the last 5
years, which considers the culture factor, its impact and how students perceive that not only
the language is taught, but also jointly the culture and the specific contexts in which it is
used. Even what is the perspective that researchers have had in making these studies and
the considerations or recommendations they can make regarding each of these teaching and
learning strategies. Research reveals that in the process of teaching and learning a second
language must be interconnected with the study of culture, since the different authors agree
that culture is an essential aspect when studying and learning a second language.
Key words: strategies, learning, teaching, culture.