2017 Teachers’ perceptions about Government Policies of Foreign Language Teaching in Cartagena de Indias, Colombia.
Abstract
This study analyzes the teachers’ perceptions in the public and private sectors around the foreign
language teaching policies in the city of Cartagena de Indias, Colombia. It is oriented to
educational research, is descriptive-analytical, since the collection of data was done through
surveys that were completed by the participants’ own perspectives and eventually analyzed. The
results of the study revealed that public school teachers seem to be more aware of these policies
while most of the private institutions teachers are unaware of these policies. Also, these results
indicated that private schools offer a better training regarding bilingual education while public
schools prepare their students for ICFES tests; This reflects the lack of feasibility in the
fulfillment the goals related to foreign language teaching policies proposed by the Ministerio de
Educación Nacional (MEN), reason for which it is recommended to adjust them.
Keywords: perceptions, bilingualism policies, bilingualism, foreign language teaching