2014 GUÍA DE DISEÑO DE MATERIALES PARA   LA ENSEÑANZA DEL INGLÉS
Abstract 
This research aims to develop a guide to design materials from 1st to 3rd grade in primary school for English teachers of Institución Educativa Corazón de María to facilitate and improvethe teaching and the learning of English, to identify and classify different types of materials and resources to teach English from grades 1st to 3rd of the institution. It is expected to overcome the shortcomings that exist when teachers do not know how to adapt materials so that the teachers can suggest valuable and effective tasks for teaching the language. Additionally, we intent to present an alternative proposal in the design guide to develop some samples of teaching materials and how to use them to serve as a guideline for teachers of English at the elementary school from 1st to 3rdgrades so that they can develop them with their students. The elaboration of the guide will help teachers become aware of making decisions, regarding the planning and organization of classroom activities, and to determine what instruments are valid, useful, relevant and appropriate to improve the teaching of English in basic primary education in the Institución Educativa Corazón de María, San José Claveriano branch. 
Key words: teaching materials, teaching and learning, selection and use of materials.