ACADEMIC APPLICATION OF ACTIVE LISTENING SKILLS FOR 3RD GRADE STUDENTS OF BILINGUAL SCHOOL OF CARTAGENA
This bilingual schoolof Cartagena was founded in August of 1952 . The school is a private, non profit institution belonging to parents and managed by the parents Association and the School Board. The school was created to crifer an integrated bilingual primary. middle, and high school education to students from Colombia, the United States and other countries.