2017 MARKETING COMO ELEMENTO DINAMIZADOR DE LA DEMANDA DE CARTAGENA DE INDIAS COMO DESTINO TURÍSTICO DE PREFERENCIA DE VIAJEROS NACIONALES
This monograph focuses in the analysis of the importance of the marketing in the promotion
of the tourism in the city of Cartagena of The Indies, implemented by CORPOTURISMO in the
period 2014-2016, illustrating the type of information that offers the natives of Cartagena of The
Indies and Colombians in general on the city and identifying his incident in the niches of markets
where they apply the strategies of promotion and positioning.
The methodology used in the accomplishment of this monograph is of type mixed, since it
rests on the qualitative part to extend the results of the quantitative information.
The methodology used in the accomplishment of this monograph is of type mixed, since it
rests on the qualitative part to extend the results of the quantitative information. In the quantitative
approach it is based on the number as information and depends on the technologies of descriptive
statistics for the analysis of information. While, in the qualitative approach, which uses the speech
as information, it depends on the analysis of content, the interview is in use as tools of compilation
of information.
There were in use tools of compilation of information as interviews. Between the obtained
results it was possible to demonstrate that the majority of the flights that come to Cartagena of The
Indies come from the cities where promotion is realized by more frequency.
Key Words:
Marketing, promotion, tourism, Cartagena, CORPOTURISMO.