2017 Propuesta de una Ruta Gastronómica en la Ciudad de Cartagena de Indias
Abstract
The gastronomic tours are tourism products that are planned to market several integrated services and this generate benefits to the communities involved in the route. This research proposes a gastronomic route design In the city of Cartagena de Indias highlighting traditional culinary secrets of it´s inhabitants with the purpose of pay tribute to the tourism development of the participating cartageneros and safeguard the culinary culture of the local cuisine.
Keywords: Tourist Route, Gastronomy, cuisine, Traditions.