2016 SERVICIO AL CLIENTE: EL ALMA DE LAS ORGANIZACIONES CONTEMPORÁNEAS
Customer service now has become a key factor in organizations, the consumer always seek to satisfy their needs and expectations. That is why companies should be at the forefront of creating strategies to provide excellent service, better than its competitors. This article seeks to college students and organizations understand the importance of customer service as the fundamental soul of enterprises, in which customers seek to have adequate and accurate response to their needs. In order for organizations to provide effective delivery of customer service that is memorable and meets the expectations the consumer wants.