2016 “UN MINUTO EN LA CABEZA DEL CONSUMIDOR”
"Consumer behavior is defined as the individual's activities aimed at the acquisition and use of a good or service including decision-making processes which precede and determine those activities, actions that the consumer takes place in the search process, purchase, use and evaluation of products they hope will serve to meet their needs.” Shiffman, Leon G. and L. Lazar Kanuk (2005); for that reason this article describes that process carries out a consumer when choosing certain product making known the background, models, approaches and different hypotheses that have developed over time, it is important for employers to know how to satisfy the desires and needs of consumers to find the best way to set goals, strategies and of course the best way to market your product or service. As sources of information were vital research done by Leon G Shiffman and Leslie Lazar Kanuk in his book "Consumer Behavior" where defined processes for consumer study giving very relevant contributions in each of its editions; further examples of how companies can influence the purchasing decision of a consumer concluding with the most important factors in the study of the behavior of a consumer as decipher and understand these behaviors serve the employer to remain in the market are provided.
PALABRAS CLAVES: Consumidor, comportamiento, marketing, estudio, enfoque, decisión y compra.