2016 NEUROMARKETING DE EMOCIONES EN LOS RESTAURANTES
This article explores the main concepts related to Neuromarketing; organizations use mainly in eating establishments, and understand the behavior of customers in the hospitality industry as one of the most sensitive, and where emotions play a role critical to sales. What better gastronomic sector to apply the Neuromarketing to increase sales?
It is very important to know where the customers will be looking when they have the letter in front? Where they stop your eyes on the letter? What is the first thing that set? If a restaurant design your letter based on Neuromarketing, to increase sales of its menu,15%