2016 ANÁLISIS ORIGEN DEL MARKETING ¿ARTE, TÉCNICA O CIENCIA?
For decades there has been an intense debate about the scientific nature of marketing. For researchers the marketing is not a science, while others think otherwise (art), we will see how it is fed from other disciplines and sciences to structure their role within the business field, which has led to the evolution of the concept, excelling in three periods, first where marketing is product-oriented , consumer second and finally today we know as human value.
Now we see that the study of consumer behavior is integrated into the overall marketing, but will not know until know the background is its origin.
Key Words: technology, consumer, sciense, art, marketing.