2016 ANÁLISIS DE LA EXCECELENCIA EN EL SERVICIO AL CLIENTE COMO FACTOR PRINCIPAL EN EL ÉXITO DE LA EMPRESAS
ABSTRACT: Within the parameters generated by the constant changes in consumer behavior, customer service satisfaction it is part of the success or failure of a company. That's why quality customer service has become a strategic decision that must implement all companies regardless of size, structure and nature of their operations. The goal of this documentary research is to show the characteristics of companies that have been focused on the satisfaction of internal and external customers.
KEY WORDS: Service, Quality, Quality of service, Customer, Customer attention, Company, Satisfaction.