2018 RESPONSABILIDAD SOCIAL DEL CONTADOR PÚBLICO EN LAS EMPRESAS DEL SECTOR FINANCIERO DE CARTAGENA DE INDIAS
ABSTRACT
This work will be aimed at describing and analyzing the concepts linked to the understanding of social responsibility and professional ethics as intrinsic factors of
10
the activity of the public accountant, from both theoretical and practical perspectives. It will be specifically delimited to the financial sector of the city of Cartagena, which represents one of the most important economic epicenters of the Caribbean region, and one of the sectors that has been the target of financial scandals and irregular management, both public and private monies, which have impacted not only the financial development of the city, but also political and social development. In this research a mixed approach was used: on the one hand, analytical-descriptive type, in terms of the description of concepts, and on the other, a qualitative approach, in terms of its application to the aforementioned context. Finally, taking as reference the sector dealt with in this work, it is stated that the accounting professional must know how to handle not only the strictly technical part of his profession, but also all those ethical criteria and axiological principles that will allow him to assume, in the presence of events negative or fraudulent, a position that contributes to safeguarding their public image, their moral integrity and the effective application of the principle of social responsibility.
KEYWORDS
Ethics, Public Accountant, Social Responsibility, Society, Financial Sector.