2017 |Propuesta De Modelo De Balance Social A La Empresa Serport S.A.
Abstract
This research work makes a general description of some existing Social Balance models, with the purpose of analyzing the most suitable model to measure the activities developed by SERPORT S.A. in matters of corporate social responsibility. It also emphasizes the importance and benefits of having a business management tool such as the Social Balance. The business environment is described in a general way, and the need to have a tool that allows to evaluate and measure the CSR with which SERPORT can demonstrate the activities it develops in terms of the subject line exposed is highlighted.
The research was carried out based on the descriptive methodology, obtaining information from various sources, such as internal documents of the company, interviews, newspapers and the review of different theses that developed the same topic. Contributing to the alternative current in social, environmental and cultural accounting (social and environmental costs).
The foregoing helped to conclude that the appropriate balance for this company is the Brazilian IBASE model, for its simple way of applying, understanding and because it is a perfect complement when presenting the company's financial information and in this way they can take decisive decisions for efficiency and effectiveness of the company's social responsibility.
Keywords
Corporate Social Responsibility (CSR), Social Balance, Environment, Stakeholders.